pred_model/requirements.txt
pandas
scikit-learn
xgboost
joblib
numpy
pymongo
python-dotenv
